+++ src/display.rs
@@ -5,6 +5,7 @@ use neovim_lib::{Neovim, NeovimApi};
 use std::fmt;
 use std::str::FromStr;
 use std::sync::{Arc, Mutex};
+use unindent::Unindent;
 
 #[derive(Clone, Debug, Ord, PartialOrd, PartialEq, Eq)]
 pub enum DisplayType {
@@ -13,6 +14,7 @@ pub enum DisplayType {
     VirtualTextOk,
     VirtualTextErr,
     Terminal,
+    TerminalWithCode,
     LongTempFloatingWindow,
     TempFloatingWindow,
     Api,
@@ -27,6 +29,7 @@ impl FromStr for DisplayType {
             "VirtualTextOk" => Ok(VirtualTextOk),
             "VirtualTextErr" => Ok(VirtualTextErr),
             "Terminal" => Ok(Terminal),
+            "TerminalWithCode" => Ok(TerminalWithCode),
             "LongTempFloatingWindow" => Ok(LongTempFloatingWindow),
             "TempFloatingWindow" => Ok(TempFloatingWindow),
             "Api" => Ok(Api),
@@ -45,6 +48,7 @@ impl fmt::Display for DisplayType {
             DisplayType::VirtualTextOk => "VirtualTextOk",
             DisplayType::VirtualTextErr => "VirtualTextErr",
             DisplayType::Terminal => "Terminal",
+            DisplayType::TerminalWithCode => "TerminalWithCode",
             DisplayType::LongTempFloatingWindow => "LongTempFloatingWindow",
             DisplayType::TempFloatingWindow => "TempFloatingWindow",
             DisplayType::Api => "Api",
@@ -59,6 +63,11 @@ pub fn display(result: Result<String, SniprunError>, nvim: Arc<Mutex<Neovim>>, d
     display_type.sort();
     display_type.dedup(); //now only uniques display types
 
+    // remove transparently incompatible/redundant displays
+    if display_type.contains(&TerminalWithCode) {
+        display_type.retain(|dt| dt != &Terminal);
+    }
+
     info!("Display type chosen: {:?}", display_type);
     for dt in display_type.iter() {
         match dt {
@@ -66,6 +75,7 @@ pub fn display(result: Result<String, SniprunError>, nvim: Arc<Mutex<Neovim>>, d
             VirtualTextOk => display_virtual_text(&result, &nvim, data, true),
             VirtualTextErr => display_virtual_text(&result, &nvim, data, false),
             Terminal => display_terminal(&result, &nvim, data),
+            TerminalWithCode => display_terminal_with_code(&result, &nvim, data),
             LongTempFloatingWindow => display_floating_window(&result, &nvim, data, true),
             TempFloatingWindow => display_floating_window(&result, &nvim, data, false),
             Api => send_api(&result, &nvim, data),
@@ -198,14 +208,54 @@ pub fn display_terminal(
     nvim: &Arc<Mutex<Neovim>>,
     data: &DataHolder,
 ) {
+    info!("data_bloc = {}", data.current_bloc);
+    let a = data.current_bloc.lines();
+    info!("length = {}", a.count());
     let res = match message {
         Ok(result) => nvim.lock().unwrap().command(&format!(
             "lua require\"sniprun.display\".write_to_term(\"{}\", true)",
-            no_output_wrap(result, data, &DisplayType::Terminal),
+            no_output_wrap(result, data, &DisplayType::Terminal).replace("\n", "\\\n"),
         )),
         Err(result) => nvim.lock().unwrap().command(&format!(
             "lua require\"sniprun.display\".write_to_term(\"{}\", false)",
-            no_output_wrap(&result.to_string(), data, &DisplayType::Terminal),
+            no_output_wrap(&result.to_string(), data, &DisplayType::Terminal).replace("\n", "\\\n"),
+        )),
+    };
+    info!("display terminal res = {:?}", res);
+}
+
+pub fn display_terminal_with_code(
+    message: &Result<String, SniprunError>,
+    nvim: &Arc<Mutex<Neovim>>,
+    data: &DataHolder,
+) {
+    let res = match message {
+        Ok(result) => nvim.lock().unwrap().command(&format!(
+            "lua require\"sniprun.display\".write_to_term(\"{}\\n{}\", true)",
+            cleanup_and_escape(
+                &format!("\n{}", &data.current_bloc)
+                    .unindent()
+                    .lines()
+                    .fold("".to_string(), |cur_bloc, line_in_bloc| {
+                        cur_bloc + "> " + line_in_bloc + "\n"
+                    })
+            )
+            .replace("\n", "\\\n"),
+            no_output_wrap(result, data, &DisplayType::TerminalWithCode).replace("\n", "\\\n"),
+        )),
+        Err(result) => nvim.lock().unwrap().command(&format!(
+            "lua require\"sniprun.display\".write_to_term(\"{}\\n{}\", false)",
+            cleanup_and_escape(
+                &format!("\n{}", &data.current_bloc)
+                    .unindent()
+                    .lines()
+                    .fold("".to_string(), |cur_bloc, line_in_bloc| {
+                        cur_bloc + "> " + line_in_bloc + "\n"
+                    })
+            )
+            .replace("\n", "\\\n"),
+            no_output_wrap(&result.to_string(), data, &DisplayType::TerminalWithCode)
+                .replace("\n", "\\\n"),
         )),
     };
     info!("display terminal res = {:?}", res);